changed: move damaris command line properties to separate file
to avoid pulling in the whole writer in DamarisKeywords.hpp
